MSS.WinMobile/MSS.WinMobile.Application.Configuration.Tests/ConfigTest.cs
using MSS.WinMobile.Application.Configuration;
using Microsoft.VisualStudio.TestTools.UnitTesting;
namespace MSS.WinMobile.Application.Configuration.Tests
{
    
    
    /// <summary>
    ///This is a test class for ConfigTest and is intended
    ///to contain all ConfigTest Unit Tests
    ///</summary>
    [TestClass()]
    public class ConfigTest
    {


        private TestContext testContextInstance;

        /// <summary>
        ///Gets or sets the test context which provides
        ///information about and functionality for the current test run.
        ///</summary>
        public TestContext TestContext
        {
            get
            {
                return testContextInstance;
            }
            set
            {
                testContextInstance = value;
            }
        }

        #region Additional test attributes
        // 
        //You can use the following additional attributes as you write your tests:
        //
        //Use ClassInitialize to run code before running the first test in the class
        //[ClassInitialize()]
        //public static void MyClassInitialize(TestContext testContext)
        //{
        //}
        //
        //Use ClassCleanup to run code after all tests in a class have run
        //[ClassCleanup()]
        //public static void MyClassCleanup()
        //{
        //}
        //
        //Use TestInitialize to run code before running each test
        //[TestInitialize()]
        //public void MyTestInitialize()
        //{
        //}
        //
        //Use TestCleanup to run code after each test has run
        //[TestCleanup()]
        //public void MyTestCleanup()
        //{
        //}
        //
        #endregion


        /// <summary>
        ///A test for Save
        ///</summary>
        [TestMethod()]
        public void SaveTest()
        {
            string configPath = string.Empty; // TODO: Initialize to an appropriate value
            Config target = new Config(configPath); // TODO: Initialize to an appropriate value
            target.Save();
            Assert.Inconclusive("A method that does not return a value cannot be verified.");
        }

        /// <summary>
        ///A test for GetSection
        ///</summary>
        [TestMethod()]
        public void GetSectionTest()
        {
            string configPath = string.Empty; // TODO: Initialize to an appropriate value
            Config target = new Config(configPath); // TODO: Initialize to an appropriate value
            string name = string.Empty; // TODO: Initialize to an appropriate value
            Section expected = null; // TODO: Initialize to an appropriate value
            Section actual;
            actual = target.GetSection(name);
            Assert.AreEqual(expected, actual);
            Assert.Inconclusive("Verify the correctness of this test method.");
        }

        /// <summary>
        ///A test for AddSection
        ///</summary>
        [TestMethod()]
        public void AddSectionTest()
        {
            string configPath = string.Empty; // TODO: Initialize to an appropriate value
            Config target = new Config(configPath); // TODO: Initialize to an appropriate value
            string name = string.Empty; // TODO: Initialize to an appropriate value
            target.AddSection(name);
            Assert.Inconclusive("A method that does not return a value cannot be verified.");
        }

        /// <summary>
        ///A test for Config Constructor
        ///</summary>
        [TestMethod()]
        public void ConfigConstructorTest()
        {
            string configPath = string.Empty; // TODO: Initialize to an appropriate value
            Config target = new Config(configPath);
            Assert.Inconclusive("TODO: Implement code to verify target");
        }
    }
}
